requirements.txt
requests==2.25.1
pytest==8.3.5
psycopg2-binary==2.9.10
redis==6.2.0
python-dotenv==1.1.0
pytest-mock==3.14.1
tenacity==9.1.2
slowapi==0.1.9
uvicorn==0.34.2
fastapi==0.115.12